Working expectMessage utility function
