+++ index.ts
@@ -1,0 +1,7 @@
+// export { expect, readline, ast}
+import expectMessage from "./expectMessage/expectMessage";
+// import readline
+// import astNodeParser
+
+// export {expect, readline, astNodeParser}
+export { expectMessage };
